refactor(plugin): node-creation module
