+++ packages/gatsby-transformer-cloudinary/node-creation/index.js
@@ -1,0 +1,7 @@
+const { createAssetNodesFromData } = require('./create-asset-nodes-from-data');
+const { createAssetNodeFromFile } = require('./create-asset-node-from-file');
+
+module.exports = {
+  createAssetNodesFromData,
+  createAssetNodeFromFile,
+};
